David Luf:
The attached patches significantly quieten the output from genapts, which in it's current form resembles the universe flying by on a bad hair day   Remember chaps, console output on Windows is slooowwwwww...  Normal service may be resumed using --verbose or -v.  I've also added a short help, obtainable with --help or -h.  I've also added a couple of extra options, --airport=abcd for just generating a particular airport, and --tile=<[we]xxx[ns]xx> for generating a 1x1 degree tile.  We currently have --chunk=<[we]xxx[ns]xx> for generating a 10x10 degree chunk, and I'd like to eventually add --tile as an option to all tools that take --chunk.

This one adds the tile option to
tgvpf.

Erik Hofman:

Some small code changes for IRIX.

+++ src/Prep/Photo/findcorners.cxx
@@ -42,7 +42,7 @@ int main( int argc, char **argv ) {
     int dx = xrwy2 - xrwy1;
     int dy = yrwy2 - yrwy1;
     cout << "dx = " << dx << " dy = " << dy << endl;
-    double distp = sqrt( dx*dx + dy*dy );
+    double distp = sqrt( double(dx*dx + dy*dy) );
     cout << "runway distance (pixels) = " << distp << endl;
 
     // find pixel resolution
@@ -52,7 +52,7 @@ int main( int argc, char **argv ) {
     cout << "pixel per feet = " << feet_to_pixel << endl;
 
     // find runway angle in image space
-    double img_angle = atan2( dy, dx );
+    double img_angle = atan2( double(dy), double(dx) );
     cout << "runway angle in image space = "
          << img_angle * SGD_RADIANS_TO_DEGREES << endl;
     // angle offset
